requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0

langchain>=0.2.14,<0.4
langchain-core>=0.2.38,<0.4
langchain-openai>=0.1.20,<0.3
langchain-experimental>=0.0.64,<0.4

pydantic>=2.7,<3
streamlit>=1.30.0
numpy>=1.26.0
tabulate>=0.9.0